+++ NEWS.md
@@ -1,6 +1,7 @@
 # 0.1-2
 1. Discard rows in closest files that have a dot (.) in the gene id column.
 2. Added background values for `gsea`.
+3. Added `rer_functional_enrichments.r` executable script.
 
 # 0.1-1
 1. Overall cleaning and NAMESPACE update