+++ NEWS.md
@@ -6,6 +6,7 @@
 5. Added `qp_genotypes` function.
 6. Added `mktest` function that works on allele table.
 7. Added `find_singletons` and `test_singleton_enrichment` functions.
+8. Function `qp_genotypes` returns empty tibble when no sample passes.
 
 # 0.1-1
 1. Overall cleaning and NAMESPACE update